UI Refactoring
The eternal task of fixing bad code continues in full force with this latest commit. As always, during this time the app may not behave as expected. Please wait for stablization updates before posting Issues to the repo.
